+++ lambdas/data_providers/sirius_provider.py
@@ -1,7 +1,7 @@
 import os
 from data_providers.authentication import SiriusAuthenticator, SiriusAuthenticationError
 from requests import Request, Session, exceptions
-from .model import Response
+from .response import Response
 from . import UpstreamExceptionError, UpstreamTimeoutError, InternalExceptionError
 import logging
